change type to long for price and sum in ImportAdd

+++ CS511_Project_QLNS/CS511_Project_QLNS/Uct_Emp_ImportAdd.cs
@@ -51,7 +51,7 @@ namespace CS511_Project_QLNS
             this.title = title;
             this.price = string.Format("{0:#,###}", int.Parse(price));
             this.quantity = quantity;
-            int s = int.Parse(quantity) * int.Parse(price);
+            long s = long.Parse(quantity) * long.Parse(price);
             this.sum = string.Format("{0:#,###}", s);
         }
     }
